Fix handling of 64 channel devices

+++ pv/data/logicsegment.cpp
@@ -643,7 +643,7 @@ void LogicSegment::append_payload_to_mipmap()
 		else if (unit_size_ == 4)
 			downsampleT<uint32_t>(src_ptr, dest_ptr, count);
 		else if (unit_size_ == 8)
-			downsampleT<uint8_t>(src_ptr, dest_ptr, count);
+			downsampleT<uint64_t>(src_ptr, dest_ptr, count);
 		else
 			downsampleGeneric(src_ptr, dest_ptr, count);
 		len_sample -= count;
